Should_Change_Security_Stamp test moved to its own class
